+++ packages/codegen/src/generate-multiple-descriptors.ts
@@ -185,6 +185,9 @@ function capitalize(value: string) {
 function generateDescriptorValuesContent(
   descriptorValues: Record<string, DescriptorValues>,
 ) {
+  const descriptorTree = mapObject(descriptorValues, (v) => v.tree)
+  const descriptorAssets = mapObject(descriptorValues, (v) => v.asset)
+
   const usages: Record<string, number> = {}
   const countUsages = (obj: Record<string, any>): void =>
     Object.entries(obj).forEach(([key, value]) => {
@@ -192,7 +195,7 @@ function generateDescriptorValuesContent(
       usages[key]++
       if (typeof value === "object") countUsages(value)
     })
-  countUsages(descriptorValues)
+  countUsages(descriptorTree)
 
   const tokens: Array<string> = []
   const tokenToIdx: Record<string, number> = {}
@@ -208,7 +211,7 @@ function generateDescriptorValuesContent(
         return [tokenToIdx[key], newValue]
       }),
     ) as T
-  const minified = mapObject(descriptorValues, minifyKeys)
+  const minified = mapObject(descriptorTree, minifyKeys)
 
   const getTreeKey = (tree: Record<string, unknown>): string =>
     Object.entries(tree)
@@ -256,7 +259,9 @@ function generateDescriptorValuesContent(
 
   const commonTrees = findCommonTrees(
     Object.keys(Object.values(minified)[0]).flatMap((type) =>
-      Object.values(minified).map((d) => d[type as keyof DescriptorValues]),
+      Object.values(minified).map(
+        (d) => d[type as keyof DescriptorValues["tree"]],
+      ),
     ),
   )
 
@@ -264,6 +269,7 @@ function generateDescriptorValuesContent(
 
   return `
     const [minified, commonTrees, tokens] = JSON.parse(\`${data}\`);
+    const assets = JSON.parse(\`${JSON.stringify(descriptorAssets)}\`)
 
     const replaceTokens = <T>(obj: Record<string | number, T>): Record<string, T> =>
       Object.fromEntries(
@@ -296,7 +302,7 @@ function generateDescriptorValuesContent(
           )
     
     const getChainDescriptors = (key: string) =>
-      unwrap(replaceTokens(minified[key]), 2)
+      ({ tree: unwrap(replaceTokens(minified[key]), 2), asset: assets[key] })
 
     ${Object.keys(descriptorValues)
       .map(