+++ packages/codegen/src/get-used-types.ts
@@ -14,6 +14,7 @@ import {
   voidEntryPointNode,
   mapEntryPointReferences,
 } from "@polkadot-api/metadata-compatibility"
+import { getAssetId } from "./generate-descriptors"
 
 /**
  * This function extracts from `metadata` all the types used from any entry
@@ -108,6 +109,8 @@ export const getUsedTypes = (
       addTypeFromEntryPoint(checksum, runtimeCallEntryPoint(method))
     }),
   )
+
+  addTypeFromLookup(getAssetId(metadata))
 
   return { types, entryPoints, checksums }
 }