+++ packages/observable-client/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Breaking
+
+- Remove `checksumBuilder` from `RuntimeContext`
+
 ## 0.3.2 - 2024-07-11
 
 ### Fixed